+++ pueue/src/daemon/network/message_handler/mod.rs
@@ -12,6 +12,7 @@ mod add;
 mod clean;
 mod edit;
 mod enqueue;
+mod env;
 mod group;
 mod kill;
 mod log;
@@ -34,6 +35,7 @@ pub fn handle_message(message: Message, state: &SharedState, settings: &Settings
         Message::Edit(editable_tasks) => edit::edit(settings, state, editable_tasks),
         Message::EditRequest(task_ids) => edit::edit_request(state, task_ids),
         Message::EditRestore(task_ids) => edit::edit_restore(state, task_ids),
+        Message::Env(message) => env::env(settings, state, message),
         Message::Enqueue(message) => enqueue::enqueue(settings, state, message),
         Message::Group(message) => group::group(settings, state, message),
         Message::Kill(message) => kill::kill(settings, state, message),