+++ pueue/tests/client/integration/mod.rs
@@ -1,6 +1,7 @@
 mod completions;
 mod configuration;
 mod edit;
+mod env;
 mod follow;
 mod group;
 mod log;